requirements.txt
cloudpickle==1.3.0
cycler==0.10.0
future==0.18.2
gym==0.17.1
imageio==2.8.0
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.2
Pillow==7.1.1
protobuf==3.11.3
pybullet==2.7.2
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.4.1
six==1.14.0
tensorboardX==2.0
torch==1.1.0
torchvision==0.3.0
tqdm==4.45.0
